Added condition to only change meta data once per file.

+++ FolderManager/MainWindow.xaml.cs
@@ -6,6 +6,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Collections.Generic;
 using TagLib;
+using System;
 
 namespace FolderManager
 {
@@ -33,10 +34,21 @@ namespace FolderManager
                 prgBar.IsEnabled = true;
 
                 var selectedPath = openFileDialog.SelectedPath;
+                int rdepth;
+
+                bool parsedSuccessfully = int.TryParse(txtRecursiveDepth.Text.Trim(), out rdepth);
+                if (!parsedSuccessfully)
+                {
+                    rdepth = 2;
+                }
 
                 Task task = new Task(() =>
                 {
-                    RecursiveTraversal(selectedPath,0);
+                    for (int i = 0; i < rdepth; i++)
+                    {
+                        RecursiveTraversal(selectedPath, 0);
+                    }
+                        
                 });
 
                 Stopwatch sw = new Stopwatch();
@@ -65,6 +77,26 @@ namespace FolderManager
             }
         }
 
+        private bool ItWasNotAlreadySortedBefore(string path)
+        {
+            try
+            {
+                var fileName = Path.GetFileNameWithoutExtension(path);
+                int fileNameInInt;
+                var validFileNameInInt = int.TryParse(fileName, out fileNameInInt);
+                if (validFileNameInInt)
+                {
+                    return fileNameInInt.ToString().Length > 2;
+                }
+            }
+            catch (Exception e)
+            {
+                throw e;
+            }
+           
+            return true;
+        }
+
         private int RecursiveTraversal(string path, int fileIndex)
         {
             //System.Threading.Thread.Sleep(5000);
@@ -79,33 +111,55 @@ namespace FolderManager
                 int maxIndex = 0;
                 foreach (var file in files)
                 {
-                    maxIndex = fileIndex + index++;
-                    var destination = Directory.GetParent(path).FullName + "\\" + maxIndex + Path.GetExtension(file);
-                    if (!System.IO.File.Exists(destination))
-                    {
-                        System.IO.File.Move(file, destination);
-                        TagLib.File tFile = TagLib.File.Create(destination);
-                        tFile.Tag.Title = "";
-                        tFile.Save();
-                    }
+                        maxIndex = fileIndex + index++;
+                        var destination = Directory.GetParent(path).FullName + "\\" + RoundIndexToTwoPlaces(maxIndex) + Path.GetExtension(file);
+                        if (!System.IO.File.Exists(destination))
+                        {
+                            System.IO.File.Move(file, destination);
+
+                            if (ItWasNotAlreadySortedBefore(file))
+                            {
+                                TagLib.File tFile = TagLib.File.Create(destination);
+                                tFile.Tag.Title = "";
+                                tFile.Save();
+                            }
+                            
+                        }
                 }
                 if (Directory.GetFiles(path).Length == 0)
                 {
                     Directory.Delete(path);
                 }
-
-                return ++maxIndex;
+                
+                 return ++maxIndex;
             }
             else
             {
                 foreach (var dir in childDirs)
                 {
                     var maxIndex = RecursiveTraversal(dir,step);
+                   
                     step = maxIndex;
                 }
             }
 
             return step;
+        }
+
+        public string[] SortedSubDirectories(string path)
+        {
+            var childDirs = Directory.GetDirectories(path);
+            childDirs = childDirs.OrderBy(d => d, comparer).ToArray();
+            return childDirs;
+        }
+
+        private static string RoundIndexToTwoPlaces(int index)
+        {
+            if (index / 10 != 0)
+            {
+                return index.ToString();
+            }
+            return string.Format("0{0}", index);
         }
 
         private class StringComparer : IComparer<string>
@@ -124,13 +178,6 @@ namespace FolderManager
                 return xresult - yresult;
 
             }
-        }
-
-        public string[] SortedSubDirectories(string path)
-        {
-            var childDirs = Directory.GetDirectories(path);
-            childDirs = childDirs.OrderBy(d => d, comparer).ToArray();
-            return childDirs;
         }
     }
 }
